feat(db) Connect DB to app services
