+++ pkg/handlers/target.go
@@ -19,9 +19,9 @@ func NewTargetHandlers(targetService interfaces.ITargetService) *TargetHandlers 
 	}
 }
 
-func (h *TargetHandlers) GetAllTargets(w http.ResponseWriter, req *http.Request) error {
+func (h *TargetHandlers) GetTargetsByTenantID(w http.ResponseWriter, req *http.Request) error {
 
-	targets, err := h.targetService.GetAllTargets("9ca3bc6c-4b78-472d-9194-62bb75d3e9fa")
+	targets, err := h.targetService.GetTargetsByTenantID("9ca3bc6c-4b78-472d-9194-62bb75d3e9fa")
 
 	if err != nil {
 		return api.WriteJSON(w, http.StatusInternalServerError, err.Error())
